hardware_language_library_extractor/requirements.txt
absl-py==0.9.0
alabaster==0.7.12
-e git://github.com/ibeltagy/allennlp@fp16_and_others#egg=allennlp
astor==0.8.1
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
blis==0.2.4
boto3==1.14.2
botocore==1.17.2
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
conllu==1.3.1
cycler==0.10.0
cymem==2.0.3
dataclasses==0.6
decorator==4.4.2
dill==0.3.0
docutils==0.15.2
editdistance==0.5.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz#egg=en_core_web_sm
filelock==3.0.12
flaky==3.7.0
Flask==1.1.2
Flask-Cors==3.0.8
ftfy==5.7
future==0.18.2
gast==0.3.3
gevent==20.6.2
google-auth==1.14.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.28.1
h5py==2.10.0
idna==2.9
imageio==2.8.0
imagesize==1.2.0
imbalanced-learn==0.6.1
importlib-metadata==1.6.1
iniconfig==1.0.0
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
jsonlines==1.2.0
jsonnet==0.16.0
jsonpickle==1.4.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lime==0.1.1.37
lxml==4.5.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.1
more-itertools==8.4.0
murmurhash==1.0.2
networkx==2.4
nltk==3.5
numpy==1.18.3
numpydoc==1.1.0
oauthlib==3.1.0
opencv-python==4.1.2.30
overrides==3.1.0
packaging==20.4
pandas==1.0.3
parsimonious==0.8.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
plac==0.9.6
pluggy==0.13.1
preshed==2.0.1
progressbar==2.5
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==1.4.2
Pygments==2.6.1
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.2
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
responses==0.10.15
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.43
scikit-image==0.15.0
scikit-learn==0.22.2.post1
scipy==1.4.1
sentencepiece==0.1.91
six==1.14.0
snowballstemmer==2.0.0
spacy<2.2,>=2.1.0
Sphinx==3.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
srsly==1.0.2
tensorboard
tensorboardX
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
thinc==7.0.8
tokenizers==0.8.0rc4
toml==0.10.1
torch==1.5.0
torchcontrib==0.0.2
tqdm==4.40.2
traitlets==4.3.3
transformers==3.0.0
Unidecode==1.1.1
urllib3==1.25.9
wasabi==0.6.0
wcwidth==0.1.9
Werkzeug==1.0.1
word2number==1.1
wrapt==1.12.1
wandb==0.9.3
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
